normalize to number of regions

+++ src/main/kotlin/util/JSONArray.kt
@@ -3,7 +3,7 @@ package util
 import java.nio.file.Files
 import java.nio.file.Path
 
-fun writeJSONArray(values: List<Int>, path: Path) {
+fun writeJSONArray(values: List<Float>, path: Path) {
     Files.createDirectories(path.parent)
     Files.newBufferedWriter(path).use { writer ->
         writer.write("[${values.joinToString(",")}]\n")
